% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{concept_detail}
\alias{concept_detail}
\title{Query the details of an Episteme API concept}
\usage{
concept_detail(country, concept)
}
\arguments{
\item{country}{The two-digit country identifier for the concept.}

\item{content}{The keyword identifier for the content of the concept (e.g. \code{"Occupation"}, \code{"Industry"}).}
}
\value{
A list of parameters (typically concept levels) associated with the concept.
}
\description{
Identify the parameters of concepts used to organise data within Emsi Episteme.
}
\examples{
concept_detail("UK", "SOC.2010")
}

