% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{concept_list}
\alias{concept_list}
\title{Query concepts used on the Episteme API}
\usage{
concept_list()
}
\value{
A full listing of concepts with description by Country and Concept.
}
\description{
Queries the Episteme API regarding concepts used to structure data and returns them in an organised list.
}
