% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{CoW}
\alias{CoW}
\title{A short-hand function to map UK ClassOfWorker dimensions around one of four options}
\usage{
CoW(option)
}
\arguments{
\item{option}{One of "E" (for Employee), "P" (for "Proprietor"), "A" (for "All" combined) or "S" (for
Employee and Proprietor separately.)}
}
\value{
The necessary mapping for ClassOfWorker for inclusion in a data pull query.
}
\description{
A short-hand function to map UK ClassOfWorker dimensions around one of four options
}
\examples{
CoW("E")
}
