% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{datapull}
\alias{datapull}
\title{Pull data from the Episteme API}
\usage{
datapull(country, content, release, constraints, metrics)
}
\arguments{
\item{country}{The two-character country code.}

\item{content}{The Emsi Episteme dataset description (e.g. "Occupation").}

\item{release}{The release or version identifier for the dataset (e.g. "2016.1").}

\item{constraints}{A list of dimensional constraints relevant to the dataset, each of which has been prepared through
\code{\link{dimmaker}} or \code{\link{CoW}}.}

\item{metrics}{A set of metrics through which to quantify the data, prepared through \code{\link{metricmaker}}.}
}
\value{
A data frame of dimensions and metrics, with dimensions classified as factors and metrics as doubles.
}
\description{
Pulls data from the Emsi Episteme API according to specified parameters and returns a prepared data frame for analysis.
}
\examples{
met1 <- data.frame(names=c("Jobs.2016","Jobs.2022"), as=c("Jobs.2016","Jobs.2022"))
met1 <- metricmaker(met1)
area1 <- data.frame(name=c("Great Britain", "Wales"), code=c("GB", "WAL"))
areadim <- dimmaker("Area", area1)
occdim <- dimmaker("Occupation", "asIdentity")
datapull("UK","Occupation","2016.1",list(CoW("A"),areadim,occsdim),met1)
}
