% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{dataset_detail}
\alias{dataset_detail}
\title{Query the details of an Episteme API dataset}
\usage{
dataset_detail(country, content, release)
}
\arguments{
\item{country}{The two-digit country identifier for the dataset.}

\item{content}{The keyword identifier for the content of the dataset (e.g. \code{"Occupation"}, \code{"Industry"}).}

\item{release}{The release or version identifier for the dataset (e.g. \code{"2016.1"}).}
}
\value{
A list with two elements: \code{Metrics} lists the Metrics available within the dataset and \code{Dimensions}
lists the dimensions requiring constraint when making a request for data. Note that \code{Metrics}
 does not include Location Quotients and Shift-Share, which are derived from the metrics here.
}
\description{
Identify the available metrics and required dimensions for a dataset held on Emsi Episteme.
}
\examples{
dataset_detail("UK","Occupation","2016.1")
}
