% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{dataset_list}
\alias{dataset_list}
\title{Query available dataset on the Episteme API}
\usage{
dataset_list()
}
\value{
A full listing of datasets with description by Country, Content and Release, to allow for filtering.
}
\description{
Queries the Episteme API regarding available datasets and returns them in an organised list.
}
