% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmsiR-package.r
\docType{package}
\name{EmsiR}
\alias{EmsiR}
\alias{EmsiR-package}
\title{EmsiR - tools for the Episteme API}
\description{
A set of tools to allow for constructing and
implementing queries for data from the labour market
intelligence available through the Emsi
Episteme API from within R.
}
