% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{EpistemeLogin}
\alias{EpistemeLogin}
\title{Set login details for the Episteme API.}
\usage{
EpistemeLogin(username, password)
}
\arguments{
\item{username}{is the episteme-username variable for Episteme.}

\item{password}{is the episteme-secret1 variable for Episteme.}
}
\value{
The variables are placed within the environment for use by other functions.
}
\description{
Sets login details for Emsi Episteme as environment variables which will be used to authenticate all queries to the API.
}

