% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{EpistemeSettings}
\alias{EpistemeSettings}
\title{Background function which is used to take Emsi Episteme login details set by \code{\link{EpistemeLogin}} and prepares the header for all queries to the API.}
\usage{
EpistemeSettings()
}
\description{
Background function which is used to take Emsi Episteme login details set by \code{\link{EpistemeLogin}} and prepares the header for all queries to the API.
}

