% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metricmaker}
\alias{metricmaker}
\title{Specify metrics for an Episteme API query}
\usage{
metricmaker(metricdf, geoparent, along)
}
\arguments{
\item{metricdf}{at minimum, a data frame with two columns: \code{name} sets out the desired names for the metrics and
\code{as} sets out the titles of the codes on Emsi Episteme. Where using derivative metrics (Openings, Location Quotients and
 Shift-Share), additional columns are required in the form of \code{metrics} to specify if they are \emph{"OP"}, \emph{"LQ"} or
 \emph{"SS"} and, for Openings and Shift-Share, a \code{base} column identifies the comparison metric for the year.}

\item{geoparent}{is required for derivative metrics, and is a geographical code identifing the parent geographical unit for analysis.}

\item{along}{is required for derivative metrics, and reflects the intended domain for analysis (e.g. "Industry" or "Occupation").}
}
\value{
A prepared data frame which will be ready for inclusion in a data pull query.
}
\description{
Takes a data frame of required metrics and necessary supporting criteria and specifies them ready for an Emsi Episteme data pull.
}
\examples{
met1 <- data.frame(names=c("Jobs.2016","Jobs.2022"), as=c("Jobs.2016","Jobs.2022"))
metricmaker(met1)
met2 <- data.frame(name=c("Jobs.2016","Jobs.2016","Jobs.2016"),as=c("Jobs16","LQ16","SS16"),metrics=c(NA,"LQ","SS"),base=c(NA,NA,"Jobs.2003"))
metricmaker(met2, "GB", "Occupation")
}

