% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{usCoW}
\alias{usCoW}
\title{A short-hand function to map US ClassOfWorker dimensions around one different options}
\usage{
usCoW(option)
}
\arguments{
\item{option}{One of "E" (for all employees, QCEW and non-QCEW),
"Q" (for QCEW Employee), "N" (for non-QCEW Employee),
"S" (for all self-employed), "SE1" (for self-employed but not extended proprietors),
"SE2" (for extended proprietors ). Finally, "A" gives all employed and self-employed,
and "S" gives all employed and self-employed broken down.}
}
\value{
The necessary mapping for ClassOfWorker for inclusion in a data pull query.
}
\description{
A short-hand function to map US ClassOfWorker dimensions around one different options
}
\examples{
CoW("E")
}
