% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{dimmaker}
\alias{dimmaker}
\title{Specify dimension for an Episteme API query}
\usage{
dimmaker(dimension, mapping)
}
\arguments{
\item{dimension}{A named dimension required by the dataset being used for analysis (e.g. "Occupation").}

\item{mapping}{This may take the form of the atomic "asIdentity" if all possible individual items are required with
no relabelling, or else a data frame with two columns: \code{name} sets labels for \code{code} which refers to the
geographic, industry, occupation or other codes used to categorise data within Episteme. Where one label is used for
multiple codes, \code{dimmaker} will merge them to form a hybrid category.}
}
\value{
A prepared list identifying the dimension and supporting mapping, ready to organise a data pull query.
}
\description{
Takes a data frame of names and codes along a dimension and specifies them ready for an Emsi Episteme data pull.
}
\examples{
mgrs <- data.frame(names=c("CEOs",
                           rep("Group of managers A",3),
                           rep("Group of managers B",4)),
                   codes=c("1115","1116","1121","1122","1123","1131","1132","1133"))
occs <- dimmaker("Occupation", mgrs)
}

